process info updates

+++ src/pipecraft-core/service_scripts/quality_filtering_paired_end_dada2.sh
@@ -84,7 +84,8 @@ printf "# Running DADA2 filterAndTrim \n"
 Rlog=$(Rscript /scripts/submodules/dada2_PE_filterAndTrim.R 2>&1)
 echo $Rlog > $output_dir/dada2_PE_filterAndTrim.log 
 wait
-printf "\n DADA2 filterAndTrim completed \n\n"
+#format R-log file
+sed -i "s/;; /\n/g" $output_dir/dada2_PE_filterAndTrim.log 
 
 
 ### Synchronizing R1 and R2 reads if $matchIDs == "true" - WORK WITH SEQKIT for matchIDs = TRUE, because sometimes DADA2 CANNOT automatically identify paired-end headers
@@ -115,10 +116,12 @@ fi
 #################################################
 ### COMPILE FINAL STATISTICS AND README FILES ###
 #################################################
-if [[ -d tempdir2 ]]; then
-    rm -rf tempdir2
+if [[ $debugger != "true" ]]; then
+    if [[ -d tempdir2 ]]; then
+        rm -rf tempdir2
+    fi
+    rm $output_dir/dada2_PE_filterAndTrim.log
 fi
-
 ### end pipe if no outputs were generated
 outfile_check=$(ls $output_dir/*.$extension 2>/dev/null | wc -l)
 if [[ $outfile_check != 0 ]]; then 
@@ -142,7 +145,7 @@ Files in 'qualFiltered_out':
     # *.rds                    = R objects for dada2.
 
 Core command -> 
-filterAndTrim(inputR1, outputR1, inputR2, outputR2, maxN = $maxN, maxEE = c($maxEE, $maxEE), truncQ = $truncQ, truncLen= c($truncLen_R1, $truncLen_R2), maxLen = $maxLen, minLen = $minLen, minQ=$minQ, rm.phix = TRUE, multithread = TRUE)
+filterAndTrim(inputR1, outputR1, inputR2, outputR2, maxN = $maxN, maxEE = c($maxEE, $maxEE), truncQ = $truncQ, truncLen= c($truncLen_R1, $truncLen_R2), maxLen = $maxLen, minLen = $minLen, minQ=$minQ, rm.phix = TRUE)
 
 Total run time was $runtime sec.
 ##################################################################
@@ -156,10 +159,11 @@ Total run time was $runtime sec.
 ########################################################" > $output_dir/README.txt
 
 #Done
-printf "\nDONE\n"
-printf "Total time: $runtime sec.\n\n"
+printf "\nDONE "
+printf "Total time: $runtime sec.\n "
 
 #variables for all services
+echo "#variables for all services: "
 echo "workingDir=$output_dir"
 echo "fileFormat=$extension"
 echo "readType=paired_end"
